add reusable DragHandle component
